+++ examples/src/main/java/picocli/examples/model/ResultHandlerWithReturnValue.java
@@ -44,8 +44,8 @@ public class ResultHandlerWithReturnValue {
 
         class Handler extends AbstractParseResultHandler<Integer> {
             public Integer handle(ParseResult pr) {
-                int count = pr.optionValue('c', 1);
-                List<File> files = pr.positionalValue(pr.matchedPositionals().get(0), Collections.<File>emptyList());
+                int count = pr.matchedOptionValue('c', 1);
+                List<File> files = pr.matchedPositionalValue(0, Collections.<File>emptyList());
                 for (File f : files) {
                     for (int i = 0; i < count; i++) {
                         System.out.println(i + " " + f.getName());