+++ RELEASE-NOTES.md
@@ -32,6 +32,7 @@ No features have been promoted in this picocli release.
 - [#283] New API: Provide `getMissing` method on MissingParameterException to get a reference to the problematic options and positional parameters. Thanks to [jcapsule](https://github.com/jcapsule) for the suggestion.
 - [#334] API Change: Renamed `ArgSpec.rawStringValues()` to `ArgSpec.stringValues()`.
 - [#342] API Change: Prefix ParseResult methods with `matched` if they return only matched options/positionals.
+- [#340] API Change: Rename `ParseResult.optionValue(String, T)` to `matchedOptionValue(String, T)`.
 - [#333] Enhancement: Added subcommand to synopsis in generated usage help. Thanks to [jcapsule](https://github.com/jcapsule) for the pull request.
 - [#323] Enhancement: Remove dependency on java.sql package: picocli should only require the java.base module when running in Java 9.
 - [#325] Enhancement: Allow custom type converter to map empty String to custom default value for empty options. Thanks to [jesselong](https://github.com/jesselong) for the suggestion.
@@ -46,14 +47,14 @@ See [3.0.0-alpha-1](https://github.com/remkop/picocli/releases/tag/v3.0.0-alpha-
 ## <a name="3.0.0-alpha-5-breaking-changes"></a> Potential breaking changes
 
 * Renamed `ArgSpec.rawStringValues()` to `ArgSpec.stringValues()`.
-* Renames `ParseResult` methods with `matched` if they return only matched options/positionals:
+* Renamed `ParseResult` methods with `matched` if they return only matched options/positionals:
     * `options` to `matchedOptions`
     * `positionalParams` to `matchedPositionals`
     * `option(char)`, `option(String)` to `matchedOption`
     * `positional(int)` to `matchedPositional`
     * `hasOption(char)`, `hasOption(String)`, `hasOption(OptionSpec)` to `hasMatchedOption`
     * `hasPositional(int)`, `hasPositional(PositionalParamSpec)` to `hasMatchedPositional`
-
+* Renamed `ParseResult.optionValue(String, T)` to `matchedOptionValue(String, T)`, and `positionalValue` to `matchedPositionalValue`.
 
 See also breaking changes for 
 [3.0.0-alpha-4](https://github.com/remkop/picocli/releases/tag/v3.0.0-alpha-4#3.0.0-alpha-4-breaking-changes),